src/mcp/requirements.txt
requests==2.32.5
traceloop-sdk==0.47.5
dotenv==0.9.9
httpx==0.28.1
langchain==1.0.3
mcp==1.22.0
fastmcp>=2.13.1
fastapi==0.120.4
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-instrumentation-langchain<0.53.0
fastmcp==2.13.1
